feat(DetailsPanel): refactor entity type icon generation and remove unused code
